+++ packages/@react-native-windows/codegen/src/index.ts
@@ -208,20 +208,34 @@ export function generate(
     methodonly,
   });
 
-  const generatorPropsH =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsH').generate;
-  const generatorPropsCPP =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsCPP').generate;
-  const generatorShadowNodeH =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeH').generate;
-  const generatorShadowNodeCPP =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeCPP').generate;
-  const generatorComponentDescriptorH =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateComponentDescriptorH').generate;
-  const generatorEventEmitterH =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateEventEmitterH').generate;
-  const generatorEventEmitterCPP =
-    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateEventEmitterCpp').generate;
+  const generatorPropsH = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GeneratePropsH',
+  )).generate;
+  const generatorPropsCPP = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GeneratePropsCPP',
+  )).generate;
+  const generatorShadowNodeH = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GenerateShadowNodeH',
+  )).generate;
+  const generatorShadowNodeCPP = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GenerateShadowNodeCPP',
+  )).generate;
+  const generatorComponentDescriptorH = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GenerateComponentDescriptorH',
+  )).generate;
+  const generatorEventEmitterH = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GenerateEventEmitterH',
+  )).generate;
+  const generatorEventEmitterCPP = require(path.resolve(
+    rncodegenPath,
+    'lib/generators/components/GenerateEventEmitterCpp',
+  )).generate;
 
   normalizeFileMap(
     generateNM2(libraryName, schema, moduleSpecName),